inspiracion/panda-etl/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.2.3"
description = "No-code ETL and data pipelines with AI and NLP"
authors = ["Pelayo <pelayo@maojolabs.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = {extras = ["all"], version = "^0.111.1"}
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
starlette = "^0.37.2"
dateparser = "^1.2.0"
requests = "^2.32.3"
google-auth = "^2.34.0"
chromadb = "^0.5.5"
openai = "~0.28.0"
groq = "^0.9.0"
schedule = "^1.2.2"
pandas = "^2.2.0"  # Or your preferred version
psycopg2-binary = "^2.9.9"
pymysql = "^1.1.1"
pymongo = "^4.8.0"
pyodbc = "^5.1.0"
oracledb = "^3.1.1"
openpyxl = "3.1.5"
pyarrow = "^20.0.0"
lxml = "4.9.3"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
ruff = "^0.3.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=xml --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "except ImportError:",
]
